+++ NcoVAppUpdate.Android/Local Notif/LocalNotif.cs
@@ -1,7 +1,4 @@
-﻿using Android.App;
-using Android.Content;
-using NcoVAppUpdate.Droid;
-using System;
+﻿using NcoVAppUpdate.Droid;
 
 [assembly: Xamarin.Forms.Dependency(typeof(LocalNotif))]
 
@@ -11,20 +8,21 @@ namespace NcoVAppUpdate.Droid
     {
         public void Alarm()
         {
-            AlarmManager manager = Application.Context.GetSystemService(Context.AlarmService) as AlarmManager;
-            Intent myIntent = new Intent(Application.Context, typeof(NotifBroadcastReceiver));
-            PendingIntent pendingIntent = PendingIntent.GetBroadcast(Application.Context, 0, myIntent, 0);
-            manager.SetRepeating(AlarmType.RtcWakeup, DateTime.Now.Millisecond, AlarmManager.IntervalDay, pendingIntent);
+            //AlarmManager manager = Application.Context.GetSystemService(Context.AlarmService) as AlarmManager;
+            //Intent myIntent = new Intent(Application.Context, typeof(NotifBroadcastReceiver));
+            //PendingIntent pendingIntent = PendingIntent.GetBroadcast(Application.Context, 0, myIntent, 0);
+            //manager.SetRepeating(AlarmType.RtcWakeup, DateTime.Now.Millisecond, AlarmManager.IntervalDay, pendingIntent);
         }
         public void CancelNotifications()
         {
-            AlarmManager manager = Application.Context.GetSystemService(Context.AlarmService) as AlarmManager;
-            Intent myIntent = new Intent(Application.Context, typeof(NotifBroadcastReceiver));
-            PendingIntent pendingIntent = PendingIntent.GetBroadcast(Application.Context, 0, myIntent, 0);
-            if (manager != null)
-            {
-                manager.Cancel(pendingIntent);
-            }
+            //AlarmManager manager = Application.Context.GetSystemService(Context.AlarmService) as AlarmManager;
+            //Intent myIntent = new Intent(Application.Context, typeof(NotifBroadcastReceiver));
+            //PendingIntent pendingIntent = PendingIntent.GetBroadcast(Application.Context, 0, myIntent, 0);
+            //if (manager != null)
+            //{
+            //    manager.Cancel(pendingIntent);
+            //}
+            //todo here
         }
     }
 }
