+++ NcoVAppUpdate/AppShell.xaml
@@ -1,49 +1,52 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
-<Shell xmlns="http://xamarin.com/schemas/2014/forms"
+<Shell x:Class="NcoVAppUpdate.AppShell"
+       xmlns="http://xamarin.com/schemas/2014/forms"
        xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
        xmlns:d="http://xamarin.com/schemas/2014/forms/design"
        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-       mc:Ignorable="d"
-       xmlns:pages="clr-namespace:NcoVAppUpdate"   
-       Shell.BackgroundColor="#212835"
-       Shell.NavBarIsVisible="True"
-       Shell.NavBarHasShadow="True"
+       xmlns:pages="clr-namespace:NcoVAppUpdate"
+       x:DataType="pages:AppShellVM"
        FlyoutIcon="hamburger.png"
-       x:Class="NcoVAppUpdate.AppShell"
-       x:DataType="pages:AppShellVM">
+       Shell.BackgroundColor="#212835"
+       Shell.NavBarHasShadow="True"
+       Shell.NavBarIsVisible="True"
+       mc:Ignorable="d">
 
     <Shell.BindingContext>
-        <pages:AppShellVM/>
+        <pages:AppShellVM />
     </Shell.BindingContext>
-    
+
     <Shell.FlyoutHeader>
-        <pages:FlyoutHeader/>
+        <pages:FlyoutHeader />
     </Shell.FlyoutHeader>
 
-    <FlyoutItem Title="International Cases" Icon="world.png">
+    <FlyoutItem Title="International Cases"
+                Icon="world.png">
         <Tab>
-            <ShellContent ContentTemplate="{DataTemplate pages:UpdatedPage}"/>
+            <ShellContent ContentTemplate="{DataTemplate pages:UpdatedPage}" />
         </Tab>
     </FlyoutItem>
 
-    <FlyoutItem Title="Contact Tracing" Icon="virus.png">
+    <FlyoutItem Title="Contact Tracing"
+                Icon="virus.png">
         <Tab>
-            <ShellContent ContentTemplate="{DataTemplate pages:ContactTracing}"/>
+            <ShellContent ContentTemplate="{DataTemplate pages:ContactTracing}" />
         </Tab>
     </FlyoutItem>
 
-    <FlyoutItem Icon="mask.png" Title="Feeling sick?">
+    <FlyoutItem Title="Feeling sick?"
+                Icon="mask.png">
         <Tab>
-            <ShellContent ContentTemplate="{DataTemplate pages:FeelingSick}"/>
+            <ShellContent ContentTemplate="{DataTemplate pages:FeelingSick}" />
         </Tab>
     </FlyoutItem>
 
-    <MenuItem Text="Developer"
+    <!--<MenuItem Command="{Binding DeveloperCredits}"
               IconImageSource="vector.png"
-              Command="{Binding DeveloperCredits}"/>
+              Text="Developer" />
 
-    <MenuItem Text="Resources credits"
+    <MenuItem Command="{Binding Credits}"
               IconImageSource="credits.png"
-              Command="{Binding Credits}"/>
+              Text="Resources credits" />-->
 
 </Shell>
