+++ NcoVAppUpdate/Models/Contact Tracing Models/DOHData.cs
@@ -1,61 +1,54 @@
-﻿using MongoDB.Bson;
-using MongoDB.Bson.Serialization.Attributes;
-using System.Collections.Generic;
-using System.ComponentModel;
-using Xamarin.Forms.Maps;
-
-namespace NcoVAppUpdate
+﻿namespace NcoVAppUpdate
 {
     //eto gawa ko, mongodb DOH OFFICIAL DATA
-    public class DOHData 
-    {
-        [BsonId]
-        public ObjectId Id { get; set; }
-        [BsonElement("region")]
-        public string Region { get; set; }
-        [BsonElement("total_cases")]
-        public int TotalCases { get; set; }
-        [BsonElement("active_cases")]
-        public int ActiveCases { get; set; }
-        [BsonElement("recovered")]
-        public int Recovered { get; set; }
-        [BsonElement("died")]
-        public int Died { get; set; }
-        [BsonElement("asymptomatic")]
-        public int Asymptomatic { get; set; }
-        [BsonElement("mild")]
-        public int Mild { get; set; }
-        [BsonElement("severe")]
-        public int Severe { get; set; }
-        [BsonElement("critical")]
-        public int Critical { get; set; }
-        [BsonElement("admitted")] // WALA TO DI NALAGAY, LALAGAY SOON
-        public int Admitted { get; set; }
-        [BsonElement("not_admitted")]
-        public int NotAdmitted { get; set; }
-        [BsonElement("admitted_for_validation")] //ETO RIN SOON
-        public int AdmittedForValidation { get; set; }
-        [BsonElement("male")]
-        public int Male { get; set; }
-        [BsonElement("female")]
-        public int Female { get; set; }
-        [BsonElement("latitude")]
-        public double Latitude { get; set; }
-        [BsonElement("longitude")]
-        public double Longitude { get; set; }
-    }
-    public class Bayanihan
-    {
-        [BsonId]
-        public ObjectId Id { get; set; }
-        [BsonElement("triage")]
-        public string Triage { get; set; }
-        [BsonElement("symptoms")]
-        public string Symptoms { get; set; }
-        [BsonElement("latitude")]
-        public double Latitude { get; set; }
-        [BsonElement("longitude")]
-        public double Longitude { get; set; }
-    }
-    
+    //public class DOHData 
+    //{
+    //    [BsonId]
+    //    public ObjectId Id { get; set; }
+    //    [BsonElement("region")]
+    //    public string Region { get; set; }
+    //    [BsonElement("total_cases")]
+    //    public int TotalCases { get; set; }
+    //    [BsonElement("active_cases")]
+    //    public int ActiveCases { get; set; }
+    //    [BsonElement("recovered")]
+    //    public int Recovered { get; set; }
+    //    [BsonElement("died")]
+    //    public int Died { get; set; }
+    //    [BsonElement("asymptomatic")]
+    //    public int Asymptomatic { get; set; }
+    //    [BsonElement("mild")]
+    //    public int Mild { get; set; }
+    //    [BsonElement("severe")]
+    //    public int Severe { get; set; }
+    //    [BsonElement("critical")]
+    //    public int Critical { get; set; }
+    //    [BsonElement("admitted")] // WALA TO DI NALAGAY, LALAGAY SOON
+    //    public int Admitted { get; set; }
+    //    [BsonElement("not_admitted")]
+    //    public int NotAdmitted { get; set; }
+    //    [BsonElement("admitted_for_validation")] //ETO RIN SOON
+    //    public int AdmittedForValidation { get; set; }
+    //    [BsonElement("male")]
+    //    public int Male { get; set; }
+    //    [BsonElement("female")]
+    //    public int Female { get; set; }
+    //    [BsonElement("latitude")]
+    //    public double Latitude { get; set; }
+    //    [BsonElement("longitude")]
+    //    public double Longitude { get; set; }
+    //}
+    //public class Bayanihan
+    //{
+    //    [BsonId]
+    //    public ObjectId Id { get; set; }
+    //    [BsonElement("triage")]
+    //    public string Triage { get; set; }
+    //    [BsonElement("symptoms")]
+    //    public string Symptoms { get; set; }
+    //    [BsonElement("latitude")]
+    //    public double Latitude { get; set; }
+    //    [BsonElement("longitude")]
+    //    public double Longitude { get; set; }
+    //}
 }
