+++ NcoVAppUpdate/NcoVAppUpdate.csproj
@@ -1,7 +1,8 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
+    <TargetFramework>netstandard2.1</TargetFramework>
+	<LangVersion>10.0</LangVersion>
     <ProduceReferenceAssembly>true</ProduceReferenceAssembly>
   </PropertyGroup>
 
@@ -11,23 +12,35 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.AppCenter.Analytics" Version="3.2.2" />
-    <PackageReference Include="Microsoft.AppCenter.Crashes" Version="3.2.2" />
-    <PackageReference Include="Microsoft.AppCenter.Distribute" Version="3.2.2" />
-    <PackageReference Include="MongoDB.Bson" Version="2.7.0" />
-    <PackageReference Include="MongoDB.Driver" Version="2.7.0" />
-    <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
-    <PackageReference Include="RestSharp" Version="106.11.4" />
-    <PackageReference Include="Syncfusion.Xamarin.SfBusyIndicator" Version="18.1.0.59" />
-    <PackageReference Include="Syncfusion.Xamarin.SfComboBox" Version="18.1.0.59" />
-    <PackageReference Include="Syncfusion.Xamarin.SfProgressBar" Version="18.1.0.59" />
-    <PackageReference Include="Syncfusion.Xamarin.SfSparkline" Version="18.1.0.59" />
-    <PackageReference Include="Xam.Plugin.SimpleBottomDrawer" Version="1.0.5" />
-    <PackageReference Include="Xamarin.Forms" Version="4.7.0.968" />
-    <PackageReference Include="Xamarin.Essentials" Version="1.5.3.2" />
-    <PackageReference Include="Xamarin.Forms.Maps" Version="4.7.0.968" />
-    <PackageReference Include="Xamarin.Forms.PancakeView" Version="1.4.2" />
-    <PackageReference Include="Xamarin.Forms.Visual.Material" Version="4.7.0.968" />
+    <None Remove="Resources\SF-Compact-Text-Medium.otf" />
+    <None Remove="Resources\SF-UI-Display-Bold.otf" />
+    <None Remove="Resources\SF-UI-Display-Regular.otf" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <EmbeddedResource Include="Resources\SF-Compact-Text-Medium.otf" />
+    <EmbeddedResource Include="Resources\SF-UI-Display-Bold.otf" />
+    <EmbeddedResource Include="Resources\SF-UI-Display-Regular.otf" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.AppCenter.Analytics" Version="4.5.3" />
+    <PackageReference Include="Microsoft.AppCenter.Crashes" Version="4.5.3" />
+    <PackageReference Include="Microsoft.AppCenter.Distribute" Version="4.5.3" />
+    <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
+    <PackageReference Include="Prism.DryIoc.Forms" Version="8.1.97" />
+    <PackageReference Include="PropertyChanged.Fody" Version="4.0.5" />
+    <PackageReference Include="Syncfusion.Xamarin.SfBusyIndicator" Version="20.3.0.50" />
+    <PackageReference Include="Syncfusion.Xamarin.SfComboBox" Version="20.3.0.50" />
+    <PackageReference Include="Syncfusion.Xamarin.SfProgressBar" Version="20.3.0.50" />
+    <PackageReference Include="Syncfusion.Xamarin.SfSparkline" Version="20.3.0.50" />
+    <PackageReference Include="Xam.Plugin.SimpleBottomDrawer" Version="1.0.11" />
+    <PackageReference Include="Xamarin.CommunityToolkit" Version="2.0.5" />
+    <PackageReference Include="Xamarin.Forms" Version="5.0.0.2515" />
+    <PackageReference Include="Xamarin.Essentials" Version="1.7.3" />
+    <PackageReference Include="Xamarin.Forms.Maps" Version="5.0.0.2515" />
+    <PackageReference Include="Xamarin.Forms.PancakeView" Version="2.3.0.759" />
+    <PackageReference Include="Xamarin.Forms.Visual.Material" Version="5.0.0.2515" />
     <PackageReference Include="Xamarin.Plugin.FilePicker" Version="2.1.41" />
   </ItemGroup>
 
@@ -38,7 +51,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <Compile Update="View\FeelingSick.xaml.cs">
+    <Compile Update="Views\FeelingSick.xaml.cs">
       <DependentUpon>FeelingSick.xaml</DependentUpon>
     </Compile>
   </ItemGroup>
@@ -47,16 +60,19 @@
     <EmbeddedResource Update="AppShell.xaml">
       <Generator>MSBuild:UpdateDesignTimeXaml</Generator>
     </EmbeddedResource>
-    <EmbeddedResource Update="View\FlyoutHeader.xaml">
+    <EmbeddedResource Update="Views\FlyoutHeader.xaml">
       <Generator>MSBuild:UpdateDesignTimeXaml</Generator>
     </EmbeddedResource>
-    <EmbeddedResource Update="View\FeelingSick.xaml">
+    <EmbeddedResource Update="Views\FeelingSick.xaml">
       <Generator>MSBuild:UpdateDesignTimeXaml</Generator>
     </EmbeddedResource>
-    <EmbeddedResource Update="View\UpdatedPage.xaml">
+    <EmbeddedResource Update="Views\HomePage.xaml">
       <Generator>MSBuild:UpdateDesignTimeXaml</Generator>
     </EmbeddedResource>
-    <EmbeddedResource Update="View\ContactTracing.xaml">
+    <EmbeddedResource Update="Views\UpdatedPage.xaml">
+      <Generator>MSBuild:UpdateDesignTimeXaml</Generator>
+    </EmbeddedResource>
+    <EmbeddedResource Update="Views\ContactTracing.xaml">
       <Generator>MSBuild:UpdateDesignTimeXaml</Generator>
     </EmbeddedResource>
   </ItemGroup>
