+++ NcoVAppUpdate/View Model/ViewModelBase.cs
@@ -1,0 +1,42 @@
+﻿using Prism.Mvvm;
+using Prism.Navigation;
+
+namespace NcoVAppUpdate.View_Model
+{
+    public class ViewModelBase : BindableBase, IInitialize, INavigationAware, IDestructible
+    {
+        protected INavigationService NavigationService { get; private set; }
+
+        private string _title;
+        public string Title
+        {
+            get => _title;
+            set => SetProperty(ref _title, value);
+        }
+
+        public ViewModelBase(INavigationService navigationService)
+        {
+            NavigationService = navigationService;
+        }
+
+        public virtual void Initialize(INavigationParameters parameters)
+        {
+
+        }
+
+        public virtual void OnNavigatedFrom(INavigationParameters parameters)
+        {
+
+        }
+
+        public virtual void OnNavigatedTo(INavigationParameters parameters)
+        {
+
+        }
+
+        public virtual void Destroy()
+        {
+
+        }
+    }
+}
