--- NcoVAppUpdate/View/FeelingSick.xaml.cs
@@ -1,14 +1,0 @@
-﻿using Xamarin.Forms;
-using Xamarin.Forms.Xaml;
-
-namespace NcoVAppUpdate
-{
-    [XamlCompilation(XamlCompilationOptions.Compile)]
-    public partial class FeelingSick : ContentPage
-    {
-        public FeelingSick()
-        {
-            InitializeComponent();
-        }
-    }
-}
