+++ NcoVAppUpdate/Views/FeelingSick.xaml.cs
@@ -1,0 +1,14 @@
+﻿using Xamarin.Forms;
+using Xamarin.Forms.Xaml;
+
+namespace NcoVAppUpdate
+{
+    [XamlCompilation(XamlCompilationOptions.Compile)]
+    public partial class FeelingSick : ContentPage
+    {
+        public FeelingSick()
+        {
+            InitializeComponent();
+        }
+    }
+}
