+++ NcoVAppUpdate/Views/FlyoutHeader.xaml
@@ -1,0 +1,14 @@
+﻿<?xml version="1.0" encoding="UTF-8"?>
+<ContentView xmlns="http://xamarin.com/schemas/2014/forms" 
+             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+             xmlns:d="http://xamarin.com/schemas/2014/forms/design"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+             mc:Ignorable="d"
+             x:Class="NcoVAppUpdate.FlyoutHeader">
+  <ContentView.Content>
+      <StackLayout HorizontalOptions="FillAndExpand" VerticalOptions="FillAndExpand">
+            <Image Source="coromaflyout2.jpg"
+                   Aspect="AspectFit"/>
+        </StackLayout>
+  </ContentView.Content>
+</ContentView>
