+++ NcoVAppUpdate/Views/HomePage.xaml.cs
@@ -1,0 +1,12 @@
+﻿using Xamarin.Forms;
+
+namespace NcoVAppUpdate.Views
+{
+    public partial class HomePage : ContentPage
+    {
+        public HomePage()
+        {
+            InitializeComponent();
+        }
+    }
+}
